refactor: distribute job routes by role
