+++ server/routes/employers.ts
@@ -8,6 +8,7 @@ import { asyncHandler } from '../utils/asyncHandler';
 import { validateBody } from '../middleware/validation';
 import { EmployerRepository, JobPostRepository } from '../repositories';
 import { storage } from '../storage';
+import { isValidTransition, canPerformAction } from '@shared/utils/jobStatus';
 
 export const employersRouter = Router();
 
@@ -132,5 +133,132 @@ employersRouter.post(
     const jobData: InsertJobPost = insertJobPostSchema.parse({ ...req.body, employerId: employer.id });
     const jobPost = await storage.createJobPost(jobData);
     res.json(jobPost);
+  })
+);
+
+// Routes moved from jobsRouter
+
+employersRouter.get(
+  '/jobs/:id',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await storage.getJobPost(jobId);
+    if (!job || (job as any).employerId !== employer.id) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    res.json(job);
+  })
+);
+
+employersRouter.put(
+  '/jobs/:id',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await storage.getJobPost(jobId);
+    if (!job) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    if (!canPerformAction('employer', job.jobStatus as any, 'edit', job.deleted)) {
+      return res.status(403).json({ message: 'Cannot edit fulfilled or deleted jobs' });
+    }
+    if ((job as any).employerId !== employer.id) {
+      return res.status(403).json({ message: 'Access denied' });
+    }
+    const updateData = insertJobPostSchema.partial().parse(req.body) as Partial<InsertJobPost>;
+    const updatedJob = await storage.updateJobPost(jobId, updateData);
+    res.json(updatedJob);
+  })
+);
+
+employersRouter.patch(
+  '/jobs/:id/fulfill',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await JobPostRepository.findById(jobId);
+    if (!job || (job as any).employerId !== employer.id) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    if (!isValidTransition(job.jobStatus as any, 'FULFILLED', job.deleted)) {
+      return res.status(400).json({ message: 'Invalid status transition' });
+    }
+    const fulfilledJob = await JobPostRepository.update(jobId, { jobStatus: 'FULFILLED' } as any);
+    res.json(fulfilledJob);
+  })
+);
+
+employersRouter.patch(
+  '/jobs/:id/activate',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await storage.getJobPost(jobId);
+    if (!job || (job as any).employerId !== employer.id) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    if (!canPerformAction('employer', job.jobStatus as any, 'activate', job.deleted)) {
+      return res.status(400).json({ message: 'Invalid status transition' });
+    }
+    if (!isValidTransition(job.jobStatus as any, 'ACTIVE', job.deleted)) {
+      return res.status(400).json({ message: 'Invalid status transition' });
+    }
+    const activatedJob = await storage.activateJob(jobId);
+    res.json(activatedJob);
+  })
+);
+
+employersRouter.patch(
+  '/jobs/:id/deactivate',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await storage.getJobPost(jobId);
+    if (!job || (job as any).employerId !== employer.id) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    if (!isValidTransition(job.jobStatus as any, 'PENDING', job.deleted)) {
+      return res.status(400).json({ message: 'Invalid status transition' });
+    }
+    const deactivatedJob = await storage.deactivateJob(jobId);
+    res.json(deactivatedJob);
+  })
+);
+
+employersRouter.get(
+  '/jobs/:id/applications',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await storage.getJobPost(jobId);
+    if (!job || (job as any).employerId !== employer.id) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    const applications = await storage.getApplicationsByJob(jobId);
+    res.json(applications);
+  })
+);
+
+employersRouter.post(
+  '/jobs/:id/clone',
+  ...requireVerifiedRole('employer'),
+  asyncHandler(async (req: any, res) => {
+    const employer = req.employer;
+    const jobId = parseInt(req.params.id);
+    const job = await storage.getJobPost(jobId);
+    if (!job || (job as any).employerId !== employer.id) {
+      return res.status(404).json({ message: 'Job not found' });
+    }
+    const jobCode = `JOB-${Date.now()}-${Math.random().toString(36).substr(2, 4).toUpperCase()}`;
+    const cloneData = { ...job, title: `Copy of ${job.title}`, employerId: employer.id, jobCode, jobStatus: 'PENDING' };
+    const clonedJob = await storage.createJobPost(cloneData);
+    res.json(clonedJob);
   })
 );
