+++ server/routes/index.ts
@@ -4,7 +4,6 @@ import express from 'express';
 import { authRouter } from './auth';
 import { candidatesRouter } from './candidates';
 import { employersRouter } from './employers';
-import { jobsRouter } from './jobs';
 import { adminRouter } from './admin';
 
 export async function registerRoutes(app: Express): Promise<Server> {
@@ -17,7 +16,6 @@ export async function registerRoutes(app: Express): Promise<Server> {
   app.use('/api/auth', authRouter);
   app.use('/api/candidates', candidatesRouter);
   app.use('/api/employers', employersRouter);
-  app.use('/api/jobs', jobsRouter);
   app.use('/api/admin', adminRouter);
 
   const httpServer = createServer(app);
